+++ Mods ChoGGi/RC Garage/Code/Script.lua
@@ -82,7 +82,7 @@ DefineClass.RCGarage = {
 	-- keeps a modifier obj to change power use
 	power_modifier = false,
 	-- show the pin info
-	pin_rollover = T{0,"<StatusUpdate>"},
+	pin_rollover = T(0,"<StatusUpdate>"),
 }
 
 function RCGarage:GameInit(...)
@@ -430,7 +430,7 @@ function OnMsg.ClassesBuilt()
 					if text_rovers:format(#context.stored_rovers) == context.status_text then
 						self:SetTitle([[Main Garage]])
 					else
-						self:SetTitle(T{0,"<StatusUpdate>"})
+						self:SetTitle(T(0,"<StatusUpdate>"))
 					end
 
 					---