+++ Mods ChoGGi/RC Bulldozer/Code/Script.lua
@@ -76,7 +76,7 @@ if BuildingTemplates.RCHarvesterBuilding then
 end
 
 local function Trans(...)
-	return _InternalTranslate(T(...))
+	return _InternalTranslate(T(pack_params(...)))
 end
 
 local idle_text = StringFormat([[Radius: %s, %s: %s]],"%s",Trans(49--[[Status--]]),Trans(949--[[Idle--]]))