+++ Expanded Cheat Menu/Changelog.md
@@ -7,6 +7,7 @@
 ### Changed:
 - Updated real time path marker to work for off-map colonists as well (ones in buildings).
 - EnumVars button to examine toolbar (only shows up when there's something to view).
+- Updated Mod Upload func with ignore_files.
 
 ### Fixed:
 - Set Colonists ages/etc didn't check if the colonist was valid (thanks tmpyemail).